+++ HISTORY.md
@@ -2,6 +2,7 @@
 ------------------
 
 - Fixed resolving of all built-in types on empty nodes.
+- Fixed invalid warning on empty lines within quoted scalars and flow collections.
 
 
 3.2.4 / 2014-12-19